reafactor: modularize machinetranslators
form each connectors subprojects
and move wiremock dependency to test fixtures
