+++ Plugins.properties
@@ -2,7 +2,6 @@ plugin=org.omegat.filters3.xml.xliff.XLIFFFilter \
     org.omegat.filters2.text.TextFilter \
     org.omegat.tokenizer.LuceneEnglishTokenizer \
     org.omegat.tokenizer.HunspellTokenizer \
-    org.omegat.core.machinetranslators.BelazarTranslate \
     org.omegat.gui.scripting.ScriptingWindow \
     org.omegat.externalfinder.ExternalFinder \
     org.omegat.gui.theme.DefaultFlatTheme \
@@ -90,19 +89,8 @@ tokenizer=org.omegat.tokenizer.DefaultTokenizer \
     org.omegat.tokenizer.LuceneSwedishTokenizer \
     org.omegat.tokenizer.LuceneThaiTokenizer \
     org.omegat.tokenizer.LuceneTurkishTokenizer
-machinetranslator=org.omegat.core.machinetranslators.Google2Translate \
-    org.omegat.core.machinetranslators.ApertiumTranslate \
-    org.omegat.core.machinetranslators.MyMemoryHumanTranslate \
-    org.omegat.core.machinetranslators.MyMemoryMachineTranslate \
-    org.omegat.core.machinetranslators.DeepLTranslate \
-    org.omegat.core.machinetranslators.IBMWatsonTranslate \
-    org.omegat.core.machinetranslators.YandexCloudTranslate
 plugin.desc.dictionary=Bundled dictionary files connectors. There are two bundled connectors. LingvoDSL, connector \
   \ for .dsl, .dsl.dz, .dsl.gz and StarDict, connector for .ifo, *.dict
-plugin.desc.machinetranslator=Bundled machine translator service connectors. There are several connectors;  \
-  Apertium Translate / Belazar Translate / DeepL Translate (only Pro user can use) /IBM Watson Translate / \
-  Microsoft Translate V2 (obsolete, only user who has v2 key can use) / MyMemory Human Translate / MyMemory Machine \
-  Translate / Google Translate v2 / Yandex Cloud Translate (obsolete, only user who already has access key can use)
 plugin.desc.misc=Bundled misc plugins
 plugin.desc.script=Script engine
 plugin.desc.guiutil=GUI extension
