+++ machinetranslators/mymemory/src/main/java/org/omegat/core/machinetranslators/AbstractMyMemoryTranslate.java
